cms/requirements.txt
asgiref==3.11.0
Django==6.0
django-cors-headers==4.9.0
django-unfold==0.72.0
django-unfold-markdown==0.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.7
PyJWT==2.10.1
sqlparse==0.5.4
tzdata==2025.2